requirements.txt
pandas
jupyter
matplotlib
ipywidgets
requests
beautifulsoup4
requests-ip-rotator
plotly
argparse
streamlit
watchdog
boto
pyarrow
pytest
s3fs
jinja2